requirements.txt
Flask==3.0.0
requests==2.31.0
beautifulsoup4==4.12.2
mecab-python3==1.0.6
anthropic==0.25.0
google-api-python-client==2.111.0
google-search-results==2.4.2
python-dotenv==1.0.0
lxml==4.9.4
unidic-lite==1.0.8
